Move home tab content to its own js module
