+++ src/index.js
@@ -1,3 +1,4 @@
+import displayContentHome from './content_home';
 import './style.scss';
 
 const contentContainer = document.getElementById('content-container');
@@ -12,17 +13,4 @@ contentContainer.innerHTML = `
                               </div>
                             `
 
-const contentHome = document.createElement('div');
-contentHome.setAttribute('id', 'content-home');
-contentHome.setAttribute('class', 'tab-content');
-
-contentHome.innerHTML = `
-                          <div id="content-wrapper-home" class="content-wrapper">
-                            <p>Fresh, daily stunned chicken picked from our own farm. Delivered to any corner in the city. Just order and follow your delivery</p>
-                            <p>We dedicate one chef each for our unique spicy Indian recipes, crispy and juicy fast served fried chicken, whole chicken rotisserie as well as Turkish Doner</p>
-                            <p>All our items are charcoal free. We use electric oven for rotiserrie, baking and roasting</p>
-                            <p>All come with our trademark Yumz cheeseNmustard&trade; sauce and a cold drink of your choice</p>
-                            <p>Choose any type and weight of chicken one day ahead with our special FancyYumzPick&reg; and pick them directly from the farm using video call </p>
-                          </div>
-                        `
-contentContainer.appendChild(contentHome);
+displayContentHome(contentContainer);
